requirements.txt
# Core dependencies
requests>=2.31.0  # HTTP requests
beautifulsoup4>=4.12.0  # HTML parsing
groq>=0.3.0  # Content processing
notion-client>=2.0.0  # Notion integration
pyyaml>=6.0  # Configuration management
python-dotenv>=1.0.0  # Environment variables management
feedparser>=6.0.0  # RSS feed parsing

# Async support
aiohttp>=3.8.0
asyncio>=3.4.3

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
lxml>=5.3.0  # Fast HTML parser

# Development dependencies
black>=22.3.0  # Code formatting
flake8>=4.0.1  # Code linting
pytest>=7.0.0  # Testing
pytest-asyncio>=0.18.0  # Async testing support
pytest-cov>=3.0.0  # Test coverage 